move speakerDetails to public folder
